modify trainner, tester and add log

+++ utils/tools.py
@@ -7,31 +7,27 @@ def info_print(head, msg):
     print(('[{:<'+str(pre_len)+'s}]{: <'+str(10-pre_len)+'s}{}').format(head, ' ', msg))
 
 
-def asset_dir():
-    new_dir = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), '../asset/'))
-    if not os.path.exists(new_dir):
-        os.makedirs(new_dir)
-        print("[Dir]       Create asset_dir in " + new_dir)
-    return new_dir
-
-
-def trainlog_dir():
-    new_dir = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), '../config/'))
+def trainlog_dir(prefix=None):
+    if prefix is not None:
+        new_dir = os.path.abspath(os.path.join(os.path.dirname(str(prefix)), '../train_log/'))
+    else:
+        new_dir = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), '../train_log/'))
     if not os.path.exists(new_dir):
         os.makedirs(new_dir)
         print("[Dir]       Create train_log in " + new_dir)
     return new_dir
 
 
-def set_dir(prefix, name):
+def set_dir(prefix, name, overwrite=False):
     new_dir = os.path.join(prefix, name)
     if not os.path.exists(new_dir):
         os.makedirs(new_dir)
         print("[Dir]       Create new_dir in " + new_dir)
     else:
-        print("[Warning]   Existing new_dir, will overwrite it.")
-        shutil.rmtree(new_dir)  # removes all the subdirectories!
-        os.makedirs(new_dir)
+        if overwrite:
+            print("[Warning]   Existing new_dir, will verwrite it.")
+            shutil.rmtree(new_dir)  # removes all the subdirectories!
+            os.makedirs(new_dir)
     return new_dir
